+++ src/card/cards/combat_constraints.rs
@@ -206,3 +206,5 @@ pub const GOBLIN_WIZARD: CardDefinitionId = CardDefinitionId(2005);
 pub const GAEAS_TOUCH: CardDefinitionId = CardDefinitionId(2006);
 pub const INTO_THE_WILDS: CardDefinitionId = CardDefinitionId(2007);
 pub const TRAUMATIZE: CardDefinitionId = CardDefinitionId(2008);
+pub const DEATHS_PRESENCE: CardDefinitionId = CardDefinitionId(2009);
+pub const AJANI_CALLER_OF_THE_PRIDE: CardDefinitionId = CardDefinitionId(2010);