+++ src/game/tests.rs
@@ -619,6 +619,7 @@ mod laces;
 mod land_and_ability_layers;
 mod land_characteristics;
 mod landwalk;
+mod life_and_death_amounts;
 mod living_artifact;
 mod looking_at_another_library;
 mod m13_more_stale;